# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Web3 backend for Sukuk (Islamic bonds) on Base Testnet, providing APIs for dApps and frontends while processing blockchain events from an indexer.",
        "title": "Sukuk POC Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.sukuk-poc.com/support",
            "email": "support@sukuk-poc.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/investments": {
            "get": {
                "description": "Get a list of all investments with optional filtering by investor address and status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments"
                ],
                "summary": "List all investments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Investor wallet address to filter by",
                        "name": "investor_address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Investment status to filter by (active, redeemed)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of investments",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.InvestmentListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/investments/investor/{address}": {
            "get": {
                "description": "Get all investments made by a specific wallet address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments"
                ],
                "summary": "Get investments by investor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Investor wallet address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Investor's investments",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.InvestmentListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid address",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/investments/sukuk/{sukukId}": {
            "get": {
                "description": "Get all investments made in a specific Sukuk series",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments"
                ],
                "summary": "Get investments by Sukuk series",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sukuk Series ID",
                        "name": "sukukId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukuk series investments",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.InvestmentListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/investments/{id}": {
            "get": {
                "description": "Get detailed information about a specific investment including Sukuk series and company data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments"
                ],
                "summary": "Get investment details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Investment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Investment details",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.InvestmentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Investment not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_kadzu_sukuk-poc-be_internal_models.Company": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Leading Indonesian sukuk issuer"
                },
                "email": {
                    "type": "string",
                    "example": "contact@sukukindonesia.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "industry": {
                    "type": "string",
                    "example": "Financial Services"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "legal_documents": {
                    "description": "JSON array of document paths",
                    "type": "string",
                    "example": "[\"doc1.pdf\", \"doc2.pdf\"]"
                },
                "logo": {
                    "description": "File path to company logo",
                    "type": "string",
                    "example": "/uploads/logos/company_1.png"
                },
                "name": {
                    "type": "string",
                    "example": "PT Sukuk Indonesia"
                },
                "sukuk_series": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.SukukSeries"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "wallet_address": {
                    "type": "string",
                    "example": "0x1234567890123456789012345678901234567890"
                },
                "website": {
                    "type": "string",
                    "example": "https://sukukindonesia.com"
                }
            }
        },
...
        "github_com_kadzu_sukuk-poc-be_internal_models.Investment": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Investment amount in IDRX",
                    "type": "string",
                    "example": "1000000000000000000000"
                },
                "block_number": {
                    "type": "integer",
                    "example": 12345678
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "investment_date": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "investor_address": {
                    "type": "string",
                    "example": "0x1234567890123456789012345678901234567890"
                },
                "investor_email": {
                    "type": "string",
                    "example": "investor@example.com"
                },
                "redemptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Redemption"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "sukuk_series": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.SukukSeries"
                        }
                    ]
                },
                "sukuk_series_id": {
                    "type": "integer",
                    "example": 1
                },
                "token_amount": {
                    "description": "Equivalent Sukuk tokens received",
                    "type": "string",
                    "example": "1000000000000000000000"
                },
                "transaction_hash": {
                    "type": "string",
                    "example": "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "yield_claims": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.YieldClaim"
                    }
                }
            }
        },
...
        "github_com_kadzu_sukuk-poc-be_internal_models.Redemption": {
            "type": "object",
            "properties": {
                "approval_notes": {
                    "description": "Company's approval/rejection notes",
                    "type": "string"
                },
                "approved_at": {
                    "type": "string"
                },
                "block_number": {
                    "description": "Set when completed",
                    "type": "integer"
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "investment": {
                    "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Investment"
                },
                "investment_id": {
                    "type": "integer"
                },
                "investor_address": {
                    "type": "string"
                },
                "redemption_amount": {
                    "description": "IDRX amount to be returned",
                    "type": "string"
                },
                "rejected_at": {
                    "type": "string"
                },
                "rejection_reason": {
                    "type": "string"
                },
                "request_reason": {
                    "description": "Why investor wants to redeem",
                    "type": "string"
                },
                "requested_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.RedemptionStatus"
                },
                "sukuk_series": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.SukukSeries"
                        }
                    ]
                },
                "sukuk_series_id": {
                    "type": "integer"
                },
                "token_amount": {
                    "description": "Amount of tokens to redeem",
                    "type": "string"
                },
                "transaction_hash": {
                    "description": "Set when completed",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_kadzu_sukuk-poc-be_internal_models.RedemptionStatus": {
            "type": "string",
            "enum": [
                "requested",
                "approved",
                "rejected",
                "completed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "RedemptionStatusRequested",
                "RedemptionStatusApproved",
                "RedemptionStatusRejected",
                "RedemptionStatusCompleted",
                "RedemptionStatusCancelled"
            ]
        },
        "github_com_kadzu_sukuk-poc-be_internal_models.SukukSeries": {
            "type": "object",
            "properties": {
                "company": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Company"
                        }
                    ]
                },
                "company_id": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Sustainable infrastructure financing sukuk"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "investments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Investment"
                    }
                },
                "is_redeemable": {
                    "description": "Can investors redeem early",
                    "type": "boolean",
                    "example": true
                },
                "legal_documents": {
                    "description": "JSON array of document paths",
                    "type": "string",
                    "example": "[\"legal1.pdf\", \"legal2.pdf\"]"
                },
                "maturity_date": {
                    "type": "string",
                    "example": "2027-12-31T00:00:00Z"
                },
                "max_investment": {
                    "description": "Maximum investment amount (optional)",
                    "type": "string",
                    "example": "100000000000000000000"
                },
                "min_investment": {
                    "description": "Minimum investment amount",
                    "type": "string",
                    "example": "1000000000000000000"
                },
                "name": {
                    "type": "string",
                    "example": "Green Sukuk Series A"
                },
                "outstanding_supply": {
                    "description": "Currently issued",
                    "type": "string",
                    "example": "500000000000000000000000"
                },
                "payment_frequency": {
                    "description": "Payments per year (quarterly = 4)",
                    "type": "integer",
                    "example": 4
                },
                "prospectus": {
                    "description": "PDF file path",
                    "type": "string",
                    "example": "/uploads/prospectus/sukuk_1.pdf"
                },
                "redemptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Redemption"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "symbol": {
                    "type": "string",
                    "example": "GSA"
                },
                "token_address": {
                    "type": "string",
                    "example": "0xabcdefabcdefabcdefabcdefabcdefabcdefabcd"
                },
                "total_supply": {
                    "description": "Use string for big numbers",
                    "type": "string",
                    "example": "1000000000000000000000000"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "yield_claims": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.YieldClaim"
                    }
                },
                "yield_rate": {
                    "description": "Annual yield rate (e.g., 0.0850 for 8.5%)",
                    "type": "number",
                    "example": 0.085
                }
            }
        },
        "github_com_kadzu_sukuk-poc-be_internal_models.YieldClaim": {
            "type": "object",
            "properties": {
                "block_number": {
                    "description": "Set when claimed",
                    "type": "integer"
                },
                "claimed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "description": "Yield claims can expire",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "investment": {
                    "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Investment"
                },
                "investment_id": {
                    "type": "integer"
                },
                "investor_address": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.YieldClaimStatus"
                },
                "sukuk_series": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.SukukSeries"
                        }
                    ]
                },
                "sukuk_series_id": {
                    "type": "integer"
                },
                "transaction_hash": {
                    "description": "Set when claimed",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "yield_amount": {
                    "description": "Yield amount in IDRX",
                    "type": "string"
                }
            }
        },
        "github_com_kadzu_sukuk-poc-be_internal_models.YieldClaimStatus": {
            "type": "string",
            "enum": [
                "pending",
                "claimed",
                "expired"
            ],
            "x-enum-varnames": [
                "YieldClaimStatusPending",
                "YieldClaimStatusClaimed",
                "YieldClaimStatusExpired"
            ]
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request data"
                }
            }
        },
...
        "internal_handlers.InvestmentListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 100
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Investment"
                    }
                }
            }
        },
        "internal_handlers.InvestmentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Investment"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for accessing protected admin endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}